requirements.txt
fastapi
pydantic-settings
python-slugify
uvicorn
